+++ app/controllers/application_controller.rb
@@ -136,4 +136,13 @@ class ApplicationController < ActionController::Base
     order = column_name.blank? ? default_order : [column_name, direction].compact.join(" ")
     order
   end
+
+  # Expects an "Uploader" type class, ex: uploader = @project.logo
+  def send_file_if_present(uploader, *args)
+    if uploader.present?
+      send_file uploader.path, *args
+    else
+      head :ok
+    end
+  end
 end