+++ app/models/admin/team_member.rb
@@ -2,9 +2,6 @@
 
 # Defines team members for team page.
 class Admin::TeamMember < ApplicationRecord
-  # Constants
-  GROUPS = %w(steering internal patient)
-
   # Uploaders
   mount_uploader :photo, PhotoUploader