+++ db/migrate/20180612214642_remove_position_default_from_team_members.rb
@@ -1,0 +1,9 @@
+class RemovePositionDefaultFromTeamMembers < ActiveRecord::Migration[5.2]
+  def up
+    change_column :admin_team_members, :position, :integer, default: nil
+  end
+
+  def down
+    change_column :admin_team_members, :position, :integer, default: 10
+  end
+end