+++ db/migrate/20180613141056_remove_group_from_admin_team_members.rb
@@ -1,0 +1,5 @@
+class RemoveGroupFromAdminTeamMembers < ActiveRecord::Migration[5.2]
+  def change
+    remove_column :admin_team_members, :group, :string
+  end
+end